Adding groups to core

+++ Apache/Ocsinventory/Server/Option/Download.pm
@@ -37,8 +37,6 @@ push @{$Apache::Ocsinventory::OPTIONS_STRUCTURE},{
 	'TYPE' => OPTION_TYPE_ASYNC
 };
 
-push @Apache::Ocsinventory::XMLParseOptForceArray, 'PACKAGE';
-
 # Default
 $Apache::Ocsinventory::OPTIONS{'OCS_OPT_DOWNLOAD'} = 0;
 $Apache::Ocsinventory::OPTIONS{'OCS_OPT_DOWNLOAD_CYCLE_LATENCY'} = 60;
@@ -109,8 +107,11 @@ sub download_pre_inventory{
 	my $data = $current_context->{'DATA'};
 	my $dbh = $current_context->{'DBI_HANDLE'};
 	my $computerId = $current_context->{'DATABASE_ID'};
-	my $result = $current_context->{'XML_ENTRY'};
+	my $result;
 	
+	unless($result = XML::Simple::XMLin( $$data, SuppressEmpty => 1, ForceArray => ['PACKAGE'] )){
+		return 1;
+	}
 	$dbh->do('DELETE FROM download_history WHERE HARDWARE_ID=(?)', {}, $computerId);
 	# Reference to the module part
